% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffectOverview.R
\name{BatchEffectOverview}
\alias{BatchEffectOverview}
\title{BatchEffectOverview}
\usage{
BatchEffectOverview(MetFlowData.before = MetFlowData1,
  MetFlowData.after = MetFlowData2, path = NULL)
}
\arguments{
\item{MetFlowData.before}{MetFlowData before normalization or integration.}

\item{path}{work directory}

\item{MetFlowData.before}{MetFlowData after normalization or integration.}
}
\value{
Give PCA score plot for QC and subject.
}
\description{
Using PCA score plot to view the batch effect.
}
\details{
Using PCA score plot to view the batch effect.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

