% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeSampleName.R
\name{ChangeSampleName}
\alias{ChangeSampleName}
\title{ChangeSampleName}
\usage{
ChangeSampleName(data = "data.csv",
  sample.information = "sample.information.csv", polarity = "positive",
  posfix = NULL, qc.has.order = FALSE, output = TRUE, path = NULL)
}
\arguments{
\item{data}{data for analysis. Default is "data.csv"}

\item{sample.information}{sample information name for analysis. First column
is sample.name, second column is injection.order, third column is
class (Subject or QC), fouth column is batch information,
fifth column is group (control or case), latter columns are other
information for sample. Please see demo data in example.}

\item{polarity}{The polarity of data. positive, negative or none, default is
positive. If data is positive mode, the name of sample should be added as
 "POS" or "NEG" as posfix. Please see the demo data in example.}

\item{posfix}{The posfix of the data. For example, "POS" or "NEG". Please
see the demo data in examples.}
}
\value{
Return a data whose name is standard name for MetFlowData.
}
\description{
Change sample name to standard name from GetWorklist.
}
\details{
Change sample name to standard name from GetWorklist.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

