% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeWorklist.R
\name{ChangeWorklist}
\alias{ChangeWorklist}
\title{ChangeWorklist}
\usage{
ChangeWorklist(date = gsub("-", "", as.character(Sys.Date())))
}
\arguments{
\item{date}{Date you want to change in you worklist. Default is today.}
}
\value{
Return a new worklist.
}
\description{
Change date in worklist from GetWorklist.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

