% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedata.R
\name{combinedata}
\alias{combinedata}
\title{combinedata}
\usage{
combinedata(...)
}
\arguments{
\item{...}{MetFlowData that you want to combine.}
}
\value{
Return a combination MetFlowData.
}
\description{
Combine several MetFlowData into one data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
