% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIntegration.R
\name{DataIntegration}
\alias{DataIntegration}
\title{DataIntegration}
\usage{
DataIntegration(MetFlowData = MetFlowData, method = "qc.mean")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{method}{Which method do you want to use?
subject.mean or qc.mean, default is qc.mean.}
}
\value{
Return a MetFlowData which has been integrated.
}
\description{
Integrate different batch datasets together.
}
\details{
Data integration is a necessary step for multiple batch dataset.
\href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:abe41368-d08d-4806-871f-3aa035d21743}{Dunn's}
method has been used in this function.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

