% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataNormalization.R
\name{DataNormalization}
\alias{DataNormalization}
\title{DataNormalization}
\usage{
DataNormalization(MetFlowData = MetFlowData, path = NULL, method = "svr",
  dimension1 = TRUE, optimization = TRUE, begin = 0.5, end = 1,
  step = 0.2, multiple = 5, threads = 2, rerun.loess = TRUE,
  rerun.svr = TRUE, peakplot = TRUE, datastyle = "tof", user = "other")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{path}{Workdirectory.}

\item{method:}{Normalization method, mean, median, total svr or loess,
default is svr. Please see the details.}

\item{dimension1:}{Remain dimension or not. Default is TRUE.}

\item{svrloess}{You can see MetNormalizer for help.}
}
\value{
The normalization results can see help for MetNormalizer.
}
\description{
Normalize dataset.
}
\details{
Data normalization is a useful method to reduce unwanted variances
in metabolomics data. The most used normalization methods is
\href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:2303e537-2864-4df5-91f0-c3ae731711af}{sample scale method}
(median, mean and total intensity). In large
scale metabolomics, QC based normalization methods are more useful.
\href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:abe41368-d08d-4806-871f-3aa035d21743}{LOESS}
and \href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:2303e537-2864-4df5-91f0-c3ae731711af}{SVR}
normalization are two most used normalization
methods.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
\seealso{
\code{\link{MetNormalizer}}
}

