% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataOverview.R
\name{DataOverview}
\alias{DataOverview}
\title{DataOverview}
\usage{
DataOverview(MetFlowData = MetFlowData, feature.distribution = TRUE,
  path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData}

\item{feature.distribution}{Draw a rt vs mz vs intensity plot or not. Default is TRUE.}
}
\value{
Data overview_RT vs mz vs intensity.pdf:  A RT vs mz vs intensity plot.

Data overview.txt: A overview information for MetFlowData.
}
\description{
Give a overview of MetFlowData.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

