% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldChange.R
\name{FoldChange}
\alias{FoldChange}
\title{FoldChange}
\usage{
FoldChange(MetFlowData = MetFlowData, to = c("case", "control"),
  ratio = "median")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{to}{Which the ratio you want? default is case/control.}

\item{ratio}{Which ratio you want to use to calculate fold change. median ot mean.}
}
\value{
MetFlowData which has been added fold change information in tags.
}
\description{
Calculate fold change for MetFlowData.
}
\details{
Calculate fold change for MetFlowData.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

