% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTestMix.R
\name{GetTestMix}
\alias{GetTestMix}
\title{GetTestMix}
\usage{
GetTestMix(data = "Test mixture.csv", test.mix.is = "test.mix.is.csv",
  test.mix.info = "test.mix.info.csv", mzerror = 15, rterror = 30,
  rt.unit.is.second = T, plot.output = TRUE, path = NULL)
}
\arguments{
\item{data}{data name for text mixture.}

\item{test.mix.is}{Internal standards information in test information.}

\item{mzerror}{mz tolerance for IS.}

\item{rterror}{rt tolerance for IS.}

\item{test.mix.data}{Test mixture data from XCMS.}
}
\description{
Test mix information.
}
\details{
Get internal standatds from data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

