% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWorklist.R
\name{GetWorklist}
\alias{GetWorklist}
\title{GetWorklist}
\usage{
GetWorklist(instrument = "Agilent", name = "worklist",
  randommethod = "no", samplenumber = NULL, replication = 1, QCstep = 8,
  conditionQCnumber = 10, testmixstep = 32, injectionfrom = 1,
  dir = "D:\\\\MassHunter\\\\Data\\\\SXT\\\\")
}
\arguments{
\item{instrument}{Which instrument you use? "Agilent" or "AB", default is "Agilent".}

\item{name}{The name of worklist.}

\item{randommethod}{Which random method you want to use? "no", "position" or "injection". Default is "no".}

\item{samplenumber}{Sample number.}

\item{replication}{Replication times.}

\item{QCstep}{QC step.}

\item{conditionQCnumber}{Condition QC number.}

\item{testmixstep}{Test mixture step.}

\item{injectionfrom}{Injection order from which? Default is 1.}

\item{dir}{Directory.}
}
\value{
New worklist.
}
\description{
Generate Worklist for data acquisition.
}
\details{
Generate Worklist for data acquisition.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

