% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMap.R
\name{HeatMap}
\alias{HeatMap}
\title{HeatMap}
\usage{
HeatMap(MetFlowData = MetFlowData, log.scale = FALSE,
  color = c("palegreen", "firebrick1", "royalblue", "yellow", "black", "cyan",
  "gray48"), variable = "all", Group = c("control", "case"),
  scale.method = "auto", show_rownames = FALSE, show_colnames = FALSE,
  path = NULL, width = 7, height = 7, border_color = NA,
  fontsize_row = 10, cluster_rows = TRUE, cluster_cols = TRUE,
  clustering_method = "ward.D", ...)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{log.scale}{log transformation or not.}

\item{color}{Color list for sample group.}

\item{variable}{"all" or "marker" for heatmap.}

\item{Group}{group for heatmap.}

\item{scale.method}{scale method.}

\item{...}{other parameters for pheatmap.}
}
\value{
A heatmap plot.
}
\description{
Heat map
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

