\name{hmdb.pathway}
\alias{hmdb.pathway}
\docType{data}
\title{
HMDB pathway.
}
\description{
HMDB pathway.
}
\usage{data(hmdb.pathway)}
\format{
  \describe{
    \item{\code{gs}}{Metabolite in each pathway.}
    \item{\code{pathwaynames}}{Pathway name.}
  }
}

\value{
HMDB pathway.
}

\examples{
data(hmdb.pathway)
}
\keyword{datasets}
