\name{hmdb.synonym}
\alias{hmdb.synonym}
\docType{data}
\title{
HMDB synonym.
}
\description{
HMDB synonym.
}
\usage{data(hmdb.synonym)}
\format{
  \describe{
    \item{\code{hmdb.id}}{HMDB ID.}
    \item{\code{synonym}}{Metabolite name.}
  }
}

\value{
HMDB synonym.
}

\examples{
data(hmdb.synonym)
}
\keyword{datasets}
