% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{ImportData}
\alias{ImportData}
\title{ImportData}
\usage{
ImportData(data = "data.csv", sample.information = "sample.information.csv",
  polarity = "positive", posfix = NULL, qc.has.order = FALSE,
  worklist.from = "manual", hasIS = "no", hasQC = "yes",
  peak.identification = "no", path = NULL)
}
\arguments{
\item{data}{Data name for analysis. Default is "data.csv". Data are csv
format from XCMS, MZmine or other software.
 Please see the demo data in example.}

\item{sample.information}{Sample information name for analysis. Default is
"sample.information.csv". Column 1 is sample.name, column 2 is
injection.order, column 3 is class ("Subject" or "QC"), column 4 is
batch information, column 5 is group ("control" or "case"), other columns
are information for sample. Please see demo data in example.}

\item{polarity}{The polarity of data, "positive", "negative"" or "none"",
default is positive.}

\item{hasQC}{The data has QC samples or not? Default is "yes".}

\item{peak.identification}{The data has identification result or not?
Default is "no".}
}
\value{
Return a standard MetProcesser dataset.
}
\description{
Import data and retrun a standard MetProcessor data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
\seealso{
\code{\link{ExportData}}
}

