\name{kegg.pathway}
\alias{kegg.pathway}
\docType{data}
\title{
KEGG pathway.
}
\description{
KEGG pathway.
}
\usage{data(kegg.pathway)}
\format{
  \describe{
    \item{\code{gs}}{Metabolite in each pathway.}
    \item{\code{pathwaynames}}{Pathway name.}
  }
}

\value{
KEGG pathway.
}

\examples{
data(kegg.pathway)
}
\keyword{datasets}
