% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2p.R
\name{m2p}
\alias{m2p}
\title{m2p}
\usage{
m2p(data = data, tags = tags, info = NULL)
}
\arguments{
\item{data}{Data (matrix) for transfomration, column correspond to samples and row correspond to metabolite.}

\item{tags}{Information (matrix) for metabolite. It must contains "Identification" which is the names of metabolites.}

\item{info}{Information (list) for samples. It must contains "normal" which are the control sample names.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
Transform metabolite infromation to pathway information using pathifier.
}
\seealso{
\code{\link[pathifier]{pathifier}}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
