% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerSelection.R
\name{MarkerSelection}
\alias{MarkerSelection}
\title{MarkerSelection}
\usage{
MarkerSelection(MetFlowData = MetFlowData, foldchange = "foldchange",
  p = "p.correct", vip = "vip", foldchange.cutoff = c(4/3, 3/4),
  p.cutoff = 0.05, vip.cutoff = 1, path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{foldchange}{which foldchange information your want to use?}

\item{p}{which p value information your want to use?}

\item{foldchange.cutoff}{The cutoff value of fold change.}

\item{p.cutoff}{The cutoff of p value.}

\item{path}{The directory you want to write results.}
}
\value{
MetFlowData: MetFlowData which has been added is.marker information into tags.

maker.information.csv: The marker information.
}
\description{
Select markers using fold change and p values.
}
\details{
Select markers using fold change and p values.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

