% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassIdentification.R
\name{MassIdentification}
\alias{MassIdentification}
\title{MassIdentification}
\usage{
MassIdentification(MetFlowData = MetFlowData, mass.tolerance = 30,
  polarity = "positive", show = 5)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{mass.tolerance}{mz tolerance.}

\item{polarity}{polarity}
}
\value{
A MetFlowData added HMDB result in tags.
}
\description{
Identify feature in HMDB according to mz.
}
\details{
\href{http://www.hmdb.ca/}{HMDB} is a metabolite database. The
database data can be loaded. In positive mode, adductions of +H, +NH4,
 +Na, +H-2H2O, +H-H2O, +K have been considered, and in negative mode, -H,
 +Cl, +CH3COO, -H-H2O, +F have been considered.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

