% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetClean.R
\name{MetClean}
\alias{MetClean}
\title{MetClean}
\usage{
MetClean(data = "data.csv", sample.information = "sample.information.csv",
  polarity = "positive", hasIS = "no", hasQC = "yes",
  obs.mv.cutoff = 0.5, var.mv.cutoff = 0.5, imputation.method = "knn",
  k = 10, rowmax = 0.5, colmax = 0.5, maxp = 1500,
  obs.zero.cutoff = 0.5, var.zero.cutoff = 0.5, method = "svr",
  threads = 2, hmdb.matching = TRUE, show = 5, mass.tolerance = 30,
  mz.tolerance = 30, rt.tolerance = 180, met.plot = TRUE, path = NULL,
  worklist.from = "manual")
}
\arguments{
\item{parameters}{All the parameters can be found in others functions.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
A whole work flow for high throughput MS based metabolomics data
cleaning.
}
\details{
The manual of MetCleaning can be found in \href{https://github.com/jaspershen/MetCleaning/blob/master/vignettes/MetCleaning.pdf}{github}
or in \href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:d8b5fcff-c725-4689-a97d-7ff106322fb6}{my library}.
}
\examples{
\dontrun{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")
data(new.group, package = "MetCleaning")
##create a folder for MetClean demo
dir.create("Demo for MetClean")
setwd("Demo for MetClean")
# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)
write.csv(new.group, "new.group.csv", row.names = FALSE)
#run MetClean
MetClean(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

