% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetCleaning-package.R
\docType{package}
\name{MetCleaning-package}
\alias{MetCleaning}
\alias{MetCleaning-package}
\title{Data cleaning and statistical analysis of large scale MS-based data.}
\usage{
MetCleaning()
}
\value{
A description of MetCleaning
}
\description{
Data cleaning and statistical analysis of large
scale MS-based data.
}
\details{
Package: MetCleaning

Type: package

Version: 0.99.0

Date: 2016-12-29

License: GPL (>= 2)
}
\author{
Xiaotao Shen

Maintainer: Xiaotao Shen shenxt@sioc.ac.cn
}

