% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetCleaning.R
\name{MetCleaning}
\alias{MetCleaning}
\title{MetCleaning}
\usage{
MetCleaning(data = "data.csv",
  sample.information = "sample.information.csv", polarity = "positive",
  hasIS = "no", hasQC = "yes", mv.filter = TRUE, obs.mv.cutoff = 0.5,
  var.mv.cutoff = 0.5, imputation.method = "knn", k = 10, rowmax = 0.5,
  colmax = 0.5, maxp = 1500, zero.filter = TRUE, obs.zero.cutoff = 0.5,
  var.zero.cutoff = 0.5, normalization = TRUE, method = "svr",
  multiple = 5, threads = 2, hmdb.matching = FALSE, show = 5,
  mass.tolerance = 30, mz.tolerance = 30, rt.tolerance = 180,
  met.plot = TRUE, path = NULL, worklist.from = "manual",
  qc.outlier.filter = TRUE, subject.outlier.filter = TRUE,
  integration = TRUE)
}
\arguments{
\item{data}{Data name for analysis. Default is "data.csv". Data are csv
format from XCMS, MZmine or other software.
 Please see the demo data in example.}

\item{sample.information}{Sample information name for analysis. Default is
"sample.information.csv". Column 1 is sample.name, column 2 is
injection.order, column 3 is class ("Subject" or "QC"), column 4 is
batch information, column 5 is group ("control" or "case"), other columns
are information for sample. Please see demo data in example.}

\item{polarity}{The polarity of data, "positive", "negative"" or "none"",
default is positive.}

\item{hasIS}{The data has IS or not? Default is "no".}

\item{hasQC}{The data has QC samples or not? Default is "yes".}

\item{mv.filter}{Filter missing values or not? Default is TRUE.}

\item{obs.mv.cutoff}{The observation MV ratio cutoff.}

\item{var.mv.cutoff}{The variable MV ratio cutoff.}

\item{imputation.method}{Which imputation method you want to use? It
contains "knn", "rf" (missForest), "mean", "median", "zero", "minium",
"bpca" (BPCA), "svd" (SVD) and "ppca" (PPCA). Default is "knn". The detial of
this method can be find in detail and reference paperes.}

\item{k}{See ?impute.knn}

\item{rowmax}{See ?impute.knn}

\item{colmax}{See ?impute.knn}

\item{maxp}{See ?impute.knn}

\item{zero.filter}{Filter zero values or not? Default is TRUE.}

\item{obs.zero.cutoff}{The observation zero ratio cutoff.}

\item{var.zero.cutoff}{The variable zero ratio cutoff.}

\item{normalization}{Data normalization or not? Default is TRUE.}

\item{method}{Normalization method, mean, median, total svr or loess,
default is svr. Please see the details.}

\item{multiple}{see ?SXTsvrNor.}

\item{threads}{Thread number.}

\item{hmdb.matching}{Default is FALSE.}

\item{show}{Default is 5.}

\item{mass.tolerance}{mz tolerance.}

\item{mz.tolerance}{mz tolerance for ms1 and ms2 data matching.}

\item{rt.tolerance}{RT tolerance for ms1 and ms2 data matching.}

\item{met.plot}{Scatter of peak.}

\item{path}{Work directory.}

\item{worklist.from}{Default is "manual".}

\item{qc.outlier.filter}{Filter QC outliers or not? Default is TRUE.}

\item{subject.outlier.filter}{Filter subject outliers or not? Default is TRUE.}

\item{integration}{Data integration or not? Default is TRUE.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
A whole work flow for high throughput MS based metabolomics data
cleaning.
}
\details{
The manual of MetCleaning can be found in \href{https://github.com/jaspershen/MetCleaning/blob/master/vignettes/MetCleaning.pdf}{github}
or in \href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:d8b5fcff-c725-4689-a97d-7ff106322fb6}{my library}.
}
\examples{
\donttest{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for MetCleaning demo
dir.create("Demo for MetCleaning")
setwd("Demo for MetCleaning")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

#run MetCleaning
MetCleaning(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
