% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetNormalizer.R
\name{MetNormalizer}
\alias{MetNormalizer}
\title{MetNormalizer}
\usage{
MetNormalizer(filename = "Metabolomics data", polarity = "none",
  minfrac.qc = 0, minfrac.sample = 0, filter = "no",
  normalization.method = "svr", optimization = TRUE, begin = 0.5,
  end = 1, step = 0.2, multiple = 5, threads = 3, rerun.loess = TRUE,
  rerun.svr = TRUE, peakplot = TRUE, datastyle = "tof",
  dimension1 = TRUE, user = "other", path = NULL)
}
\arguments{
\item{polarity}{The polarity of data, default is "none".}

\item{normalization.method}{svr: SVR normalization; loess: LOESS normalization; all: SVR&LOESS.Default is svr.}

\item{optimization}{Parameter optimization or not?.Default is TRUE.}

\item{multiple}{If multiple = 1, the svr will be built using injection order. If multiple >= 2, the svr is built using top mutiple peaks correlated peaks. For tof data, default is 5, for mrm data, default is 1.}

\item{peakplot}{Draw peakplot for each peak or nor? Default is TRUE.}

\item{dimension1}{The data after normalization is given dimension or not. Defaulte is TRUE.}

\item{path}{Directory}

\item{begin_end_step}{begin The begin of span to optimize. end: The end of span to optimize. step: The step for begin to end. begin should be set bigger if the QC number is small.}
}
\value{
peakplot: A folder contains all the peaks plot before and sfter SVR normalization.

data svr nor.csv: A csv data after SVR normalization.

normalization file: The intermediate data.

RSD compare plot.jpeg: A figure show the comparing of RSDs of peaks before and after SVR normalization.

RSD distribution.csv: RSD distributions of data before and after SVR normalization.

RSD distribution.jpeg: A figure show the RSD distributions before and after SVR normalization.

rsd.change.csv: A table show the RSDs of peaks before and after SVR normalization.
}
\description{
Normalize data using different normalization methods.
}
\details{
Normalize data using different normalization methods.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

