% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetPre.R
\name{MetPre}
\alias{MetPre}
\title{MetPre}
\usage{
MetPre(data = "data.csv", sample.information = "sample.information.csv",
  polarity = "positive", hasIS = "no", hasQC = "yes",
  obs.mv.cutoff = 0.5, var.mv.cutoff = 0.5, imputation.method = "knn",
  k = 10, rowmax = 0.5, colmax = 0.5, maxp = 1500,
  obs.zero.cutoff = 0.5, var.zero.cutoff = 0.5, method = "svr",
  threads = 2, hmdb.matching = TRUE, show = 5, mass.tolerance = 30,
  mz.tolerance = 30, rt.tolerance = 180, met.plot = TRUE, path = NULL,
  worklist.from = "manual")
}
\arguments{
\item{parameters}{All the parameters can be found in others functions.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
A whole work flow for high throughput MS based metabolomics data
cleaning.
}
\examples{
\dontrun{
#load the demo data
data(data, package = "MetProcesser")
data(sample.information, package = "MetProcesser")
data(ms2_50_300, package = "MetProcesser")
data(ms2_50_1200, package = "MetProcesser")
data(ms2_290_600, package = "MetProcesser")
data(ms2_590_900, package = "MetProcesser")
##create a folder for MetPre demo
dir.create("Demo for MetPre")
setwd("Demo for MetPre")
dir.create("peak identification")
path <- file.path(getwd(), "peak identification")
dir.create(path)
# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information , "sample.information.csv", row.names = FALSE)
write.csv(sample.information , "new.group.csv", row.names = FALSE)
write.csv(ms2_50_300, file.path(path, "ms2_50_300.csv"), row.names = FALSE)
write.csv(ms2_50_1200, file.path(path, "ms2_50_1200.csv"), row.names = FALSE)
write.csv(ms2_290_600, file.path(path, "ms2_290_600.csv"), row.names = FALSE)
write.csv(ms2_590_900, file.path(path, "ms2_590_900.csv"), row.names = FALSE)
#run MetPre
MetPre(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

