% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRMdataMerge.R
\name{MRMdataMerge}
\alias{MRMdataMerge}
\title{MRMdataMerge}
\usage{
MRMdataMerge(..., path = NULL, merge.method = "name",
  merge.term.index = c(1), tags.index = c(1:17),
  tags.same.in.datasets = c(1:4),
  tags.should.be.mean.in.datasets = c(11:13),
  tags.should.be.remain.in.datasets = NULL, mz.tolerance = 15,
  rt.tolerance = 15)
}
\arguments{
\item{...}{The MRM data to be merged.}

\item{path}{Directory for reading data and outputing results.}

\item{merge.term.index}{Use which term to merge datasets?}

\item{tags.index}{Which columna are tags in your data?}

\item{tags.same.in.datasets}{Which columna are same in you datasets?And those information only remain one in merging.}

\item{tags.should.be.mean.in.datasets}{Which columna should be averaged in the mergin}
}
\description{
Merge different MRM batch datasets into one dataset.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

