% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVImputation.R
\name{MVimputation}
\alias{MVimputation}
\title{MVimputation}
\usage{
MVimputation(MetFlowData = MetFlowData, imputation.method = "knn", k = 10,
  rowmax = 0.5, colmax = 0.8, maxp = 1500, rng.seed = 362436069,
  maxiter = 10, ntree = 100, decreasing = FALSE, replace = TRUE,
  classwt = NULL, cutoff = NULL, strata = NULL, sampsize = NULL,
  nodesize = NULL, maxnodes = NULL, xtrue = NA, parallelize = "no",
  nPcs = 2, maxSteps = 100, threshold = 1e-04)
}
\arguments{
\item{MetFlowData:}{MetFlowData.}

\item{imputation.method:}{Which imputation method you want to use? It
contains "knn", "rf" (missForest), "mean", "median", "zero", "minium",
"bpca" (BPCA), "svd" (SVD) and "ppca" (PPCA). Default is "knn". The detial of
this method can be find in detail and reference paperes.}

\item{knn}{You can see them from help of impute package: impute.knn.}

\item{missForest}{You can see them from help of missForest package: missForest.}

\item{BPCA_PPCA_SVD}{You can see them from help of pcaMethods: bpca, ppca, svd.}
}
\value{
Return a MetFlowData whose MVs have been imputated.
}
\description{
Impute MV in data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
\references{
The MV imputation in metabolomics data can see in
\href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:c9d05d0f-e945-43d0-bb4a-50ea0f90338e}{Guida's} paper.
}
\seealso{
The MV imputation methods can see in
\code{\link[impute]{impute.knn}}, \code{\link[missForest]{missForest}},
\code{\link[pcaMethods]{bpca}}, \code{\link[pcaMethods]{ppca}} and
\code{\link[pcaMethods]{svf}}.
}

