\name{my.pathway}
\alias{my.pathway}
\docType{data}
\title{
KEGG and HMDB pathway.
}
\description{
KEGG and HMDB pathway.
}
\usage{data(my.pathway)}
\format{
  \describe{
    \item{\code{gs}}{Metabolite in each pathway.}
    \item{\code{pathwaynames}}{Pathway name.}
  }
}

\value{
KEGG and HMDB pathway.
}

\examples{
data(my.pathway)
}
\keyword{datasets}
