% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZoverview.R
\name{MZoverview}
\alias{MZoverview}
\title{MZOverview}
\usage{
MZoverview(MetFlowData = MetFlowData, path = NULL, what = "mv")
}
\arguments{
\item{path}{Work Directory.}

\item{what}{Missing value ("mv") or zero values ("zero").}

\item{MetFlowData:}{MetFlowData.}
}
\value{
Batch i feature MV/zero ratio.csv: Batch i feature MV/zero ratio
distribution.

Batch i QC feature MV/zero ratio.pdf: Batch i QC feature MV.zero
ratio distribution.

Batch i Subject feature MV/zero ratio.pdf: Batch i Subject feature
 MV/zero ratio distribution.

Sample MV/zero distribution.csv: Sample MV/zero ratio distribution.
}
\description{
Evaluate the missing values or zero value information
in MetFlowData.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

