% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAanalysis.R
\name{PCAanalysis}
\alias{PCAanalysis}
\title{PCAanalysis}
\usage{
PCAanalysis(MetFlowData = MetFlowData, QC = FALSE, log.scale = FALSE,
  scale.method = "auto", path = NULL, color = c("palegreen", "royalblue",
  "firebrick1", "tan1", "deepskyblue", "cyan", "gray48", "chocolate4",
  "darkmagenta", "indianred1"), shape = c(17, 19, 15, 18, 2, 8, 11, 13, 12,
  14), cex.lab = 1.3, cex.axis = 1.3, cex = 1.3, cex.text = 1,
  width = 7, height = 7, text = FALSE, ellipse = TRUE, xlim1 = NULL,
  ylim1 = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{QC}{Use qc data for PCA analyis or not? Default is FALSE.}

\item{log.scale}{Which transformation methd you want to use? 2, 10  or "e",
defaulit is FALSE mean don't transformation.}

\item{scale.method}{Which scale methd you want to use? "auto" or "pareto",
defaulit is "auto".}

\item{path}{Work Directory}

\item{color}{Color for different class.}

\item{shape}{Shape for different class. please see the help of par(pch).}

\item{text}{Add text in PCA score plot or not? Deefault is FALSE.}

\item{ellipse}{Add ellipse in PCA score plot or not? Deefault is TRUE.}

\item{xlim1_ylim1}{The x and y axis limitation. Default is NULL.}
}
\value{
PCA score plot.
}
\description{
PCA analysis for MetFlowData.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

