% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSanalysis.R
\name{PLSanalysis}
\alias{PLSanalysis}
\title{PLSanalysis}
\usage{
PLSanalysis(MetFlowData = MetFlowData, log.scale = 10,
  scalemethod = "auto", plsmethod = "plsreg", path = NULL, width = 7,
  height = 7, text = FALSE, ellipse = TRUE, color = c("palegreen",
  "firebrick1", "royalblue", "yellow", "black", "cyan", "gray48"),
  shape = c(17, 19, 15, 18, 2, 8, 11), cexa = 1, xlim1 = NULL,
  ylim1 = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{path}{Work directory.}

\item{text}{Add text in PCA score plot or not? Deefault is FALSE.}

\item{ellipse}{Add ellipse in PCA score plot or not? Deefault is TRUE.}

\item{color}{Color for different class.}

\item{shape}{Shape for different class.}

\item{scale.method}{Whihch scale methd you want to use? "auto" or "pareto",
defaulit is "auto".}

\item{xlim1_ylim1}{The x and y axis limitation. Default is NULL.}
}
\value{
PLS score plot: PLS score plot.

permutation test plot: Permutation test plot.
}
\description{
PLS analysis for MetFlowData.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

