% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSpermutation.R
\name{PLSpermutation}
\alias{PLSpermutation}
\title{PLSpermutation}
\usage{
PLSpermutation(data = NULL, log.scale = FALSE, info = NULL,
  repeats = 200, ncomp = 3, scalemethod = "auto", path = NULL)
}
\arguments{
\item{data}{data for PLS permutation test.}

\item{log.scale}{log transformation.}

\item{info}{infomation for group.}

\item{path}{Work directory.}

\item{scale.method}{Whihch scale methd you want to use? "auto" or "pareto", defaulit is "auto".}
}
\value{
permutation test plot: Permutation test plot.
}
\description{
Permutation test for PLS analysis.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

