% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintMethodsForMyClass.R
\name{print.MetFlowData}
\alias{print.MetFlowData}
\title{PrintMethod}
\usage{
\method{print}{MetFlowData}(MetFlowData)
}
\arguments{
\item{myclass}{The name of class object in MetProcesser.}
}
\value{
Print some information of class object in your screen.
}
\description{
Print method for class in MetProcesser.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
\seealso{
\code{\link{print}}
}

