% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCOutlierFinder.R
\name{QCOutlierFinder}
\alias{QCOutlierFinder}
\title{QCOutlierFinder}
\usage{
QCOutlierFinder(MetFlowData = MetFlowData, CI = 0.95, path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{CI}{Confidence interva.}

\item{Directory}{Work directory.}
}
\value{
QCOutlierFinderData contains MetFlowData and QC outlier index (obs.remove).
}
\description{
Using PCA to find QC outliers.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

