% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReChangeGroup.R
\name{ReChangeGroup}
\alias{ReChangeGroup}
\title{ReChangeGroup}
\usage{
ReChangeGroup(MetFlowData = MetFlowData, new.group = "new.group.csv")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{new.groupNew}{Group information name. Default is new.group.csv. It can
be use the sample.information which is only changed the group column.}
}
\value{
Return a standard MetProcesser data which is changed the
group informatio.
}
\description{
Change the group information in MetFlowData for statistical
analysis.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

