% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDfilter.R
\name{RSDfilter}
\alias{RSDfilter}
\title{RSDfilter}
\usage{
RSDfilter(MetFlowData, rsd.cutoff = 30)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{rsd.cutoff}{The cutoff value of RSD. Default is 30. It means that for
a feature, if its RSD larger than 30\%, it will be removed form the dataset.}
}
\value{
Return a MetFlowData whose features have been removed.
}
\description{
Filter features according to RSD in QC samples.
}
\details{
\href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:abe41368-d08d-4806-871f-3aa035d21743}{Dunn}
recommen the cutoff value of RSD should be set as 20\% in LC-MS
data and 30\% in GC-MS data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

