% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDoverview.R
\name{RSDoverview}
\alias{RSDoverview}
\title{RSDoverview}
\usage{
RSDoverview(MetFlowData.before = MetFlowData1,
  MetFlowData.after = MetFlowData2, path = NULL)
}
\arguments{
\item{MetFlowData.before}{MetFlowData.before.}

\item{MetFlowData.after}{MetFlowData.after.}

\item{path}{Work directory.}
}
\value{
RSD comparation.
}
\description{
Evaluate the RSD information before and after processing.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

