% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitBatch.R
\name{SplitBatch}
\alias{SplitBatch}
\title{SplitBatch}
\usage{
SplitBatch(MetFlowData = MetFlowData)
}
\arguments{
\item{MetFlowData}{MetFlowData.}
}
\value{
Return a data (list), subject, qc, subject.info and qc.info.
}
\description{
Split MetFlowData accoding to different batch.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

