% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubjectOutlierFinder.R
\name{SubjectOutlierFinder}
\alias{SubjectOutlierFinder}
\title{SubjectOutlierFinder}
\usage{
SubjectOutlierFinder(MetFlowData = MetFlowData, CI = 0.95, path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{CI}{Confidence interva.}

\item{path}{Work directory.}
}
\value{
SubjectOutlierFinderData contains MetFlowData and subject outlier index (obs.remove).
}
\description{
Using PCA to find subject outliers.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
\seealso{
\code{\link{SubjectOutlierFilter}}
}

