% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXTMinifrac.R
\name{SXTMinifrac}
\alias{SXTMinifrac}
\title{SXTMinifrac}
\usage{
SXTMinifrac(..., filter.item = "mv", filter.rule = "intersect",
  minifrac.variable = 0.5, minifrac.observation = 0.5)
}
\arguments{
\item{...}{Datases you want to remove features and samples.}

\item{filter.item}{The item you want to filter. MV (missing values) or zero. Default is MV.}

\item{minifrac.variable}{The cutoff for variable. Default is 0.5.}

\item{filter.rule:}{For feature filtering, which rule you want to use, intersect or union. Default is intersect.}

\item{minifrac.observation:}{The cutoff for observation. Default is 0.5.}
}
\value{
Return a SXTMinifracData.
}
\description{
Remove feature or samples according to the MV or zero ratio.
}
\examples{
\dontrun{
## Generate
data1 <- matrix(1:20, ncol = 4)
data2 <- matrix(21:40, ncol = 4)
## Give MV in to data1 and data2
data1[sample(1:20,5)] <- NA
data2[sample(1:20,5)] <- NA
## Run SXTMinifrac
SXTMinifracData <- SXTMinifrac(data1, data2,
                              filter.item = "MV",
                              filter.rule = "intersect",
                              minifrac.variable = 0.4,
                              minifrac.observation = 0.4)
attributes(SXTMinifracData)
print(SXTMinifracData)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

