% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXTMTmatch.R
\name{SXTMTmatch}
\alias{SXTMTmatch}
\title{SXTMTmatch}
\usage{
SXTMTmatch(data1, data2, mz.tolerance = 25, rt.tolerance = 180)
}
\arguments{
\item{data1}{First data for matching, first column must be mz and seconod column must be rt.}

\item{mz.tolerance}{mz tolerance for ms1 and ms2 data matching.}

\item{rt.tolerance}{RT tolerance for ms1 and ms2 data matching.}

\item{datad2}{Second data for matching, first column must be mz and seconod column must be rt.}
}
\value{
Return a result which give the matching result of data1 and database.
}
\description{
Match two data according to mz and RT.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

