% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXTMVImputation.R
\name{SXTMVimputation}
\alias{SXTMVimputation}
\title{SXTMVimputation}
\usage{
SXTMVimputation(data, method = "knn", k = 10, rowmax = 0.5,
  colmax = 0.8, maxp = 1500, rng.seed = 362436069, maxiter = 10,
  ntree = 100, decreasing = FALSE, mtry = floor(sqrt(nrow(data))),
  replace = TRUE, classwt = NULL, cutoff = NULL, strata = NULL,
  sampsize = NULL, nodesize = NULL, maxnodes = NULL, xtrue = NA,
  parallelize = "no", nPcs = 2, maxSteps = 100, threshold = 1e-04)
}
\arguments{
\item{data}{Data to be imputated, column is sample and row is feature.}

\item{imputation.method}{Which imputation method you want to use?}

\item{knn}{You can see them from help of impute package: impute.knn.}

\item{missForest}{You can see them from help of missForest package: missForest.}

\item{BPCA_PPCA_SVD}{You can see them from help of pcaMethods: bpca, ppca, svd.}
}
\value{
Return a data whose MVs have been imputated.
}
\description{
Impute MV in data.
}
\details{
Impute MV in data..
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
\seealso{
\code{\link{MVimputation}}
}

