% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXTpca.R
\name{SXTpca}
\alias{SXTpca}
\title{SXTpca}
\usage{
SXTpca(subject = NULL, qc = NULL, info = NULL, QC = FALSE,
  scale.method = "auto", path = NULL)
}
\arguments{
\item{subject}{Subject data for PCA analysis.}

\item{qc}{QC data for PCA analysis.}

\item{info}{Group information for subject data.}

\item{QC}{Use qc data for PCA analyis or not? Default is FALSE.}

\item{scale.method}{Whihch scale methd you want to use? "auto" or "pareto", defaulit is "auto".}

\item{path}{Work directory.}
}
\value{
Return a SXTpcaData.
}
\description{
PCA analysis.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

