% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXTpcaFindOutlier.R
\name{SXTpcaFindOutlier}
\alias{SXTpcaFindOutlier}
\title{SXTpcaFindOutlier}
\usage{
SXTpcaFindOutlier(SXTpcaData = SXTpcaData, CI = 0.95, output.plot = TRUE,
  plot.name = "PCA score plot for outliers", path = NULL)
}
\arguments{
\item{SXTpcaData}{SXTpcaData subject for outliers finding.}

\item{CI}{Confidence interval.}

\item{output.plot}{Output PCA score plot or not.Default is TRUE.}

\item{plot.name}{The name of score plot.}

\item{path}{Directory you want to output results.}
}
\value{
It return a outlier index.
}
\description{
Find which samples are in outside of 95\% confidence interval in PCA score plot.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

