% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXTpcaPlot.R
\name{SXTpcaPlot}
\alias{SXTpcaPlot}
\title{SXTpcaPlot}
\usage{
SXTpcaPlot(SXTpcaData = SXTpcaData, plot = "score",
  loading.plot.name = "PCA loading", score.plot.name = "PCA score",
  color = c("palegreen", "royalblue", "firebrick1", "tan1", "deepskyblue",
  "cyan", "gray48", "chocolate4", "darkmagenta", "indianred1"), shape = c(17,
  19, 15, 18, 2, 8, 11, 13, 12, 14), cexlab = 1.3, cexaxis = 1.3,
  cexa = 1.3, cextext = 1, width = 7, height = 7, text = FALSE,
  ellipse = FALSE, path = NULL)
}
\arguments{
\item{SXTpcaData}{SXTpcaPlot subject for drawing plots.}

\item{plot}{Which class of plot you want to draw, "score" or "loading", default is "score".}

\item{loading.plot.name}{The name of loading plot.}

\item{score.plot.name}{The name of score plot.}

\item{color}{The colors you want to use in different classs.}

\item{shape}{The shapes you want to use in different classes.}

\item{text}{Add text or not? Default is FALSE.}

\item{ellipse}{Draw ellipse or not?}

\item{path}{Directory you want to output results.}
}
\value{
PCA score and loading plot.
}
\description{
Draw PCA score or loading plot for SXTpcaData.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

