% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivariateTest.R
\name{UnivariateTest}
\alias{UnivariateTest}
\title{UnivariateTest}
\usage{
UnivariateTest(MetFlowData = MetFlowData, test.method = "t",
  adjust.method = "fdr", log.scale = FALSE, class = c("control", "case"))
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{test.method}{Which test you want to use? "t" means stutent t test and "wilcox" mean wilcoxon test.}

\item{adjust.method}{p value correction method. See p.adjust function.}
}
\value{
MetFlowData which has been added p and AUC information in tags.
}
\description{
Calculate p value and AUC for each feature.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
\seealso{
The details of univariate test can be found in \code{\link[stats]{t.test}},
\code{\link[stats]{p.adjust}} and \code{\link[stats]{wilcox.test}}.
}

