% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{VolcanoPlot}
\usage{
VolcanoPlot(MetFlowData = MetFlowData, x = "foldchange", y = "p",
  z = "vip", col = c("black", "firebrick1"), foldchange.cutoff = c(4/3,
  3/4), p.cutoff = 0.05, vip.cutoff = 1, path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{col}{Colour for markers and non-markers.}

\item{path}{The directory you want to write results.}

\item{foldchange_p_foldchange.cutoff_p.cutoff}{See MarkerSeletion.}
}
\value{
volcano plot.
}
\description{
Draw volcano plot.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

