% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffectOverview.R
\name{BatchEffectOverview}
\alias{BatchEffectOverview}
\title{BatchEffectOverview}
\usage{
BatchEffectOverview(MetFlowData.before, MetFlowData.after, path = NULL)
}
\arguments{
\item{MetFlowData.before}{MetFlowData before normalization or integration.}

\item{MetFlowData.after}{MetFlowData after normalization or integration.}

\item{path}{work directory}
}
\value{
Give PCA score plot for QC and subject.
}
\description{
Using PCA score plot to view the batch effect.
}
\examples{
\dontrun{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# MetCleaning process
MetCleaning(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)

#run BatchEffectOverview
BatchEffectOverview(met.data.after.pre ,
                    met.data.after.pre,
                    path = "Demo for BatchEffectOverview")
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
