% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportData.R
\name{ExportData}
\alias{ExportData}
\title{ExportData}
\usage{
ExportData(MetFlowData = MetFlowData, data.name = "data_new",
  subject.info.name = "subject.info", qc.info.name = "qc.info",
  path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{data.name}{The name of the data you want to output. Default is
"data_new".}

\item{subject.info.name}{The name for subject information you want to
output. Default is "subject.info".}

\item{qc.info.name}{The name for QC information you want to output.
Default is "qc.info".}

\item{path}{Work directory.}
}
\value{
Write csv data.
}
\description{
Export MetFlowData as csv.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
\seealso{
\code{\link{ImportData}}
}

