% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportData.R
\name{ExportData}
\alias{ExportData}
\title{ExportData}
\usage{
ExportData(MetFlowData, data.name = "data_new",
  subject.info.name = "subject.info", qc.info.name = "qc.info",
  path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{data.name}{The name of the data you want to output. Default is
"data_new".}

\item{subject.info.name}{The name for subject information you want to
output. Default is "subject.info".}

\item{qc.info.name}{The name for QC information you want to output.
Default is "qc.info".}

\item{path}{Work directory.}
}
\value{
Write csv data.
}
\description{
Export MetFlowData as csv.
}
\examples{
\dontrun{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# MetCleaning process
MetCleaning(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
## run
ExportData(MetFlowData = met.data.after.pre, path = "Demo for ExportData")
}
}
\seealso{
\code{\link{ImportData}}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
