% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldChange.R
\name{FoldChange}
\alias{FoldChange}
\title{FoldChange}
\usage{
FoldChange(MetFlowData, to = c("case", "control"), ratio = "median")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{to}{Which the ratio you want? default is case/control.}

\item{ratio}{Which ratio you want to use to calculate fold change. median ot mean.}
}
\value{
MetFlowData which has been added fold change information in tags.
}
\description{
Calculate fold change for MetFlowData.
}
\examples{
\dontrun{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# MetCleaning process
MetCleaning(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
## run
new.met.data <- FoldChange(MetFlowData = met.data.after.pre,
                          to = c("1", "0"))
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
