% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIS.R
\name{GetIS}
\alias{GetIS}
\title{GetIS}
\usage{
GetIS(MetFlowData, mzerror = 15, rterror = 30, rt.unit.is.second = T,
  IS = "IS.csv", plot.output = TRUE, path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData for internal standatd getting.}

\item{mzerror}{mz tolerance for IS.}

\item{rterror}{rt tolerance for IS.}

\item{rt.unit.is.second}{RT's unit is second or not? Default is T.}

\item{IS}{IS data name for reading.}

\item{plot.output}{Output plot or not? Default is TRUE.}

\item{path}{Directory for outputing results.}
}
\value{
newIS: A new IS data for IS.
}
\description{
Get internal standatds from data.
}
\details{
Get internal standatds from data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

