% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IScorrection.R
\name{IScorrction}
\alias{IScorrction}
\title{IScorrection}
\usage{
IScorrction(MetFlowData, mzerror = 15, rterror = 30,
  rt.unit.is.second = TRUE, IS = "IS.csv", plot.output = TRUE,
  path = NULL, dimension1 = TRUE)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{mzerror}{mz tolerance.}

\item{rterror}{rt tolerance.}

\item{rt.unit.is.second}{RT's unit is second or not? Default is TRUE.}

\item{IS}{IS name.}

\item{plot.output}{Output plot or not? Default is TRUE.}

\item{path}{work directory.}

\item{dimension1}{Keep dimension or not?}
}
\value{
Return a new MetFlowData after IS correction.
}
\description{
Use internal standards to correct data.
}
\details{
Use internal standards to correct data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

