% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerShow.R
\name{MarkerShow}
\alias{MarkerShow}
\title{MarkerShow}
\usage{
MarkerShow(MetFlowData, beeswarm = TRUE, path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{beeswarm}{Do you want draw beeswarm on the boxplot? Deafult is TRUE.}

\item{path}{The directory you want to write results.}
}
\value{
Box plot of markers.
}
\description{
Draw boxplot for markers.
}
\examples{
\dontrun{
## load the demo data
data(met.data.after.pre, package = "MetCleaning")

##create a folder for demo
dir.create("Demo")
setwd("Demo")

## run
MarkerShow(met.data.after.pre)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
