% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolitePlot.R
\name{MetabolitePlot}
\alias{MetabolitePlot}
\title{MetabolitePlot}
\usage{
MetabolitePlot(MetFlowData.before, MetFlowData.after, path = NULL,
  figure = "jpeg")
}
\arguments{
\item{MetFlowData.before}{MetFlowData before normalization or integration.}

\item{MetFlowData.after}{MetFlowData after normalization or integration.}

\item{path}{Work directory.}

\item{figure}{Figure type you want to draw. jpeg ot pdf, default is jpeg.}
}
\value{
Return the metabolite plot before and after processing.
}
\description{
Give scatter plot for each feature.
}
\examples{
\dontrun{
## load the demo data
data(met.data.after.pre, package = "MetCleaning")

##create a folder for demo
dir.create("Demo")
setwd("Demo")

## run
MetabolitePlot(met.data.after.pre, met.data.after.pre)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

