% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolitePlot.R
\name{MetabolitePlot}
\alias{MetabolitePlot}
\title{MetabolitePlot}
\usage{
MetabolitePlot(MetFlowData.before, MetFlowData.after, path = NULL,
  figure = "jpeg")
}
\arguments{
\item{MetFlowData.before}{MetFlowData before normalization or integration.}

\item{MetFlowData.after}{MetFlowData after normalization or integration.}

\item{path}{Work directory.}

\item{figure}{Figure type you want to draw. jpeg ot pdf, default is jpeg.}
}
\value{
Return the metabolite plot before and after processing.
}
\description{
Give scatter plot for each feature.
}
\examples{
\dontrun{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# MetCleaning process
MetCleaning(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
## run
MetabolitePlot(met.data.after.pre,
              met.data.after.pre,
              path = "Demo for MetabolitePlot")
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
