% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetStat.R
\name{MetStat}
\alias{MetStat}
\title{MetStat}
\usage{
MetStat(MetFlowData = MetFlowData, new.group = TRUE, rsd.cutoff = 30,
  log.scale = FALSE, QC = TRUE, scale.method = "auto",
  plsmethod = "plsr", fc = TRUE, to = c("case", "control"),
  ratio = "median", test.method = "t", adjust.method = "fdr",
  foldchange = "foldchange", p = "p", vip = "vip",
  foldchange.cutoff = c(4/3, 3/4), p.cutoff = 0.05, vip.cutoff = 0,
  x = "foldchange", y = "p", z = "vip", col = c("black", "firebrick1"),
  variable = "all", Group = c("control", "case"), beeswarm = TRUE,
  pls.analysis = TRUE, pca.analysis = TRUE, uni.test = TRUE,
  path = NULL)
}
\arguments{
\item{parameters}{All the parameters can be found in others functions.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
A whole work flow for high throughput MS based metabolomics
data statistical analysis.
}
\examples{
\dontrun{
## load the demo data
data(met.data.after.pre, package = "MetCleaning")
data(new.group, package = "MetCleaning")
##create a folder for MetStat demo
dir.create("Demo for MetStat")
setwd("Demo for MetStat")
## export the demo data as csv
write.csv(new.group, "new.group.csv", row.names = FALSE)
## run MetStat
MetStat(MetFlowData = met.data.after.pre, new.group = TRUE)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

