% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZfinder.R
\name{MZfinder}
\alias{MZfinder}
\title{MZfinder}
\usage{
MZfinder(MetFlowData, obs.per.cutoff = 0.5, var.per.cutoff = 0.5,
  what = "mv", path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{obs.per.cutoff}{The observation MV/zero ratio cutoff.}

\item{var.per.cutoff}{The variable MV/zero ratio cutoff.}

\item{what}{Filter missing values ("mv") or zero values ("zero")?}

\item{path}{Work directory.}
}
\value{
Return a MZfinderData contains MetFlowData,feature.remove, qc.remove and subject.remove.
}
\description{
Find feature and sample outliers according to MV/zero ratio.
}
\examples{
\dontrun{
## load the demo data
data(met.data.after.pre, package = "MetCleaning")
## run
new.data <- MZfinder(met.data.after.pre)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

