% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZoverview.R
\name{MZoverview}
\alias{MZoverview}
\title{MZOverview}
\usage{
MZoverview(MetFlowData, path = NULL, what = "mv")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{path}{Work Directory.}

\item{what}{Missing value ("mv") or zero values ("zero").}
}
\value{
Batch i feature MV/zero ratio.csv: Batch i feature MV/zero ratio
distribution.

Batch i QC feature MV/zero ratio.pdf: Batch i QC feature MV.zero
ratio distribution.

Batch i Subject feature MV/zero ratio.pdf: Batch i Subject feature
 MV/zero ratio distribution.

Sample MV/zero distribution.csv: Sample MV/zero ratio distribution.
}
\description{
Evaluate the missing values or zero value information
in MetFlowData.
}
\examples{
\dontrun{
## load the demo data
data(met.data.after.pre, package = "MetCleaning")

##create a folder for demo
dir.create("Demo")
setwd("Demo")

## run
MZoverview(met.data.after.pre, what = "zero")
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

