% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakIdentification.R
\name{PeakIdentification}
\alias{PeakIdentification}
\title{PeakIdentification}
\usage{
PeakIdentification(MetFlowData, path = "peak identification",
  identification.information.from = "XCMS", peak.number = NULL,
  mz.tolerance = 30, rt.tolerance = 180, re.match = TRUE)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{path}{Work directory.}

\item{identification.information.from}{Default is "XCMS".}

\item{peak.number}{Peak number.}

\item{mz.tolerance}{mz tolerance for ms1 and ms2 data matching.}

\item{rt.tolerance}{RT tolerance for ms1 and ms2 data matching.}

\item{re.match}{Default is TRUE.}
}
\value{
Return a MetFlowData which has been added peak identification
information into tags.
}
\description{
Identify the features using other identification information
 data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

