% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakIdentification.R
\name{PeakIdentification}
\alias{PeakIdentification}
\title{PeakIdentification}
\usage{
PeakIdentification(MetFlowData, path = "peak identification",
  identification.information.from = "XCMS", peak.number = NULL,
  mz.tolerance = 30, rt.tolerance = 180, re.match = TRUE)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{path}{Work directory.}

\item{identification.information.from}{Default is "XCMS".}

\item{peak.number}{Peak number.}

\item{mz.tolerance}{mz tolerance for ms1 and ms2 data matching.}

\item{rt.tolerance}{RT tolerance for ms1 and ms2 data matching.}

\item{re.match}{Default is TRUE.}
}
\value{
Return a MetFlowData which has been added peak identification
information into tags.
}
\description{
Identify the features using other identification information
 data.
}
\examples{
\donttest{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# MetCleaning process
MetCleaning(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)

data(ms2_1, package = "MetCleaning")
dir.create("peak identification")
write.csv(ms2_1, "peak identification/ms2_1.csv", row.names = FALSE)
## run
new.met.data <- PeakIdentification(met.data.after.pre)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
