% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintMethodsForMyClass.R
\name{print.MetFlowData}
\alias{print.MetFlowData}
\title{Print.MetFlowData}
\usage{
\method{print}{MetFlowData}(x, ...)
}
\arguments{
\item{x}{MetFlowData.}

\item{...}{Other parameters for print.}
}
\value{
Print some information of class object in your screen.
}
\description{
Print method for class in MetProcesser.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
\seealso{
\code{\link{print}}
}

