% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintMethodsForMyClass.R
\name{print.MetFlowData}
\alias{print.MetFlowData}
\title{Print.MetFlowData}
\usage{
\method{print}{MetFlowData}(x, ...)
}
\arguments{
\item{x}{MetFlowData.}

\item{...}{Other parameters for print.}
}
\value{
Print some information of class object in your screen.
}
\description{
Print method for class in MetProcesser.
}
\examples{
\dontrun{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# MetCleaning process
MetCleaning(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)

print(met.data.after.pre)
}
}
\seealso{
\code{\link{print}}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
