% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCOutlierFilter.R
\name{QCOutlierFilter}
\alias{QCOutlierFilter}
\title{QCOutlierFilter}
\usage{
QCOutlierFilter(MetFlowData, CI = 0.95, path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{CI}{confidence interval.}

\item{path}{Work directory.}
}
\value{
MetFlowData whose qc outliers have been removed.
}
\description{
Using PCA to find QC outliers.
}
\examples{
\dontrun{
## load the demo data
data(met.data.after.pre, package = "MetCleaning")

##create a folder for demo
dir.create("Demo")
setwd("Demo")

## run
new.met.data <- QCOutlierFilter(met.data.after.pre)
}
}
\seealso{
\code{\link{SubjectOutlierFilter}}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
