% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCOutlierFilter.R
\name{QCOutlierFilter}
\alias{QCOutlierFilter}
\title{QCOutlierFilter}
\usage{
QCOutlierFilter(MetFlowData, CI = 0.95, path = ".")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{CI}{confidence interval.}

\item{path}{Work directory.}
}
\value{
MetFlowData whose qc outliers have been removed.
}
\description{
Using PCA to find QC outliers.
}
\examples{
\donttest{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# MetCleaning process
MetCleaning(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)


## run
new.met.data <- QCOutlierFilter(met.data.after.pre)
}
}
\seealso{
\code{\link{SubjectOutlierFilter}}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
