% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCOutlierFinder.R
\name{QCOutlierFinder}
\alias{QCOutlierFinder}
\title{QCOutlierFinder}
\usage{
QCOutlierFinder(MetFlowData, CI = 0.95, path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{CI}{Confidence interva.}

\item{path}{Work directory.}
}
\value{
QCOutlierFinderData contains MetFlowData and QC outlier index (obs.remove).
}
\description{
Using PCA to find QC outliers.
}
\examples{
\dontrun{
## load the demo data
data(met.data.after.pre, package = "MetCleaning")

##create a folder for demo
dir.create("Demo")
setwd("Demo")

## run
new.met.data <- QCOutlierFinder(met.data.after.pre)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

