% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDoverview.R
\name{RSDoverview}
\alias{RSDoverview}
\title{RSDoverview}
\usage{
RSDoverview(MetFlowData.before, MetFlowData.after, path = NULL)
}
\arguments{
\item{MetFlowData.before}{MetFlowData.before.}

\item{MetFlowData.after}{MetFlowData.after.}

\item{path}{Work directory.}
}
\value{
RSD comparation.
}
\description{
Evaluate the RSD information before and after processing.
}
\examples{
\dontrun{
## load the demo data
data(met.data.after.pre, package = "MetCleaning")

##create a folder for demo
dir.create("Demo")
setwd("Demo")

## run
RSDoverview(met.data.after.pre, met.data.after.pre)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

