% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDoverview.R
\name{RSDoverview}
\alias{RSDoverview}
\title{RSDoverview}
\usage{
RSDoverview(MetFlowData.before, MetFlowData.after, path = NULL)
}
\arguments{
\item{MetFlowData.before}{MetFlowData.before.}

\item{MetFlowData.after}{MetFlowData.after.}

\item{path}{Work directory.}
}
\value{
RSD comparation.
}
\description{
Evaluate the RSD information before and after processing.
}
\examples{
\dontrun{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# MetCleaning process
MetCleaning(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
## run
RSDoverview(met.data.after.pre, met.data.after.pre, path = "Demo for RSDoverview")
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
