% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitBatch.R
\name{SplitBatch}
\alias{SplitBatch}
\title{SplitBatch}
\usage{
SplitBatch(MetFlowData)
}
\arguments{
\item{MetFlowData}{MetFlowData.}
}
\value{
Return a data (list), subject, qc, subject.info and qc.info.
}
\description{
Split MetFlowData accoding to different batch.
}
\examples{
\dontrun{
## load the demo data
data(met.data.after.pre, package = "MetCleaning")
## run
new.data <- SplitBatch(met.data.after.pre)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
