% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitBatch.R
\name{SplitBatch}
\alias{SplitBatch}
\title{SplitBatch}
\usage{
SplitBatch(MetFlowData)
}
\arguments{
\item{MetFlowData}{MetFlowData.}
}
\value{
Return a data (list), subject, qc, subject.info and qc.info.
}
\description{
Split MetFlowData accoding to different batch.
}
\examples{
\dontrun{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# MetCleaning process
MetCleaning(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
new.data <- SplitBatch(met.data.after.pre)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
