% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubjectOutlierFilter.R
\name{SubjectOutlierFilter}
\alias{SubjectOutlierFilter}
\title{SubjectOutlierFilter}
\usage{
SubjectOutlierFilter(MetFlowData, CI = 0.95, path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{CI}{Confidence interva.}

\item{path}{Work directory.}
}
\value{
MetFlowData whose subject outliers have been removed.
}
\description{
Using PCA to filter subject outliers.
}
\examples{
\dontrun{
## load the demo data
data(met.data.after.pre, package = "MetCleaning")

##create a folder for demo
dir.create("Demo")
setwd("Demo")

## run
new.met.data <- SubjectOutlierFilter(met.data.after.pre)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
\seealso{
\code{\link{QCOutlierFilter}}
}

