% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXTMinifrac.R
\name{SXTMinifrac}
\alias{SXTMinifrac}
\title{SXTMinifrac}
\usage{
SXTMinifrac(..., filter.item = "mv", filter.rule = "intersect",
  minifrac.variable = 0.5, minifrac.observation = 0.5)
}
\arguments{
\item{...}{Datases you want to remove features and samples.}

\item{filter.item}{The item you want to filter. MV (missing values) or zero. Default is MV.}

\item{filter.rule}{For feature filtering, which rule you want to use, intersect or union. Default is intersect.}

\item{minifrac.variable}{The cutoff for variable. Default is 0.5.}

\item{minifrac.observation}{The cutoff for observation. Default is 0.5.}
}
\value{
Return a SXTMinifracData.
}
\description{
Remove feature or samples according to the MV or zero ratio.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
