% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXTMVImputation.R
\name{SXTMVimputation}
\alias{SXTMVimputation}
\title{SXTMVimputation}
\usage{
SXTMVimputation(data, method = "knn", k = 10, rowmax = 0.5,
  colmax = 0.8, maxp = 1500, rng.seed = 362436069, maxiter = 10,
  ntree = 100, decreasing = FALSE, mtry = floor(sqrt(nrow(data))),
  replace = TRUE, classwt = NULL, cutoff = NULL, strata = NULL,
  sampsize = NULL, nodesize = NULL, maxnodes = NULL, xtrue = NA,
  parallelize = "no", nPcs = 2, maxSteps = 100, threshold = 1e-04)
}
\arguments{
\item{data}{Data to be imputated, column is sample and row is feature.}

\item{method}{Which imputation method you want to use?}

\item{k}{See ?impute.knn}

\item{rowmax}{See ?impute.knn}

\item{colmax}{See ?impute.knn}

\item{maxp}{See ?impute.knn}

\item{rng.seed}{See ?impute.knn}

\item{maxiter}{See ?missForest}

\item{ntree}{See ?missForest}

\item{decreasing}{See ?missForest}

\item{mtry}{See ?missForest}

\item{replace}{See ?missForest}

\item{classwt}{See ?missForest}

\item{cutoff}{See ?missForest}

\item{strata}{See ?missForest}

\item{sampsize}{See ?missForest}

\item{nodesize}{See ?missForest}

\item{maxnodes}{See ?missForest}

\item{xtrue}{See ?missForest}

\item{parallelize}{See ?missForest}

\item{nPcs}{See ?bpca}

\item{maxSteps}{See ?bpca}

\item{threshold}{See ?bpca}
}
\value{
Return a data whose MVs have been imputated.
}
\description{
Impute MV in data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
\seealso{
\code{\link{MVimputation}}
}

