% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivariateTest.R
\name{UnivariateTest}
\alias{UnivariateTest}
\title{UnivariateTest}
\usage{
UnivariateTest(MetFlowData, test.method = "t", alternative = "two.sided",
  paired = FALSE, adjust.method = "fdr", log.scale = FALSE,
  class = c("control", "case"))
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{test.method}{Which test you want to use? "t" means stutent t test and "wilcox" mean wilcoxon test.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".
You can specify just the initial letter. See ?t.test and ?wilcox.test.}

\item{paired}{a logical indicating whether you want a paired test.
See ?t.test and ?wilcox.test.}

\item{adjust.method}{p value correction method. See p.adjust function.}

\item{log.scale}{Data transformation method, defaulst is FALSE.}

\item{class}{Class used to do test.}
}
\value{
MetFlowData which has been added p and AUC information in tags.
}
\description{
Calculate p value and AUC for each feature.
}
\examples{
\dontrun{
## load the demo data
data(met.data.after.pre, package = "MetCleaning")

##create a folder for demo
dir.create("Demo")
setwd("Demo")

## run
new.met.data <- UnivariateTest(met.data.after.pre)
}
}
\seealso{
The details of univariate test can be found in \code{\link[stats]{t.test}},
\code{\link[stats]{p.adjust}} and \code{\link[stats]{wilcox.test}}.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
