% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{VolcanoPlot}
\usage{
VolcanoPlot(MetFlowData, x = "foldchange", y = "p", z = "vip",
  col = c("grey", "tomato"), foldchange.cutoff = c(4/3, 3/4),
  p.cutoff = 0.05, vip.cutoff = 1, path = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{x}{x axis factor.}

\item{y}{y axis factor.}

\item{z}{z axis factor.}

\item{col}{Colour for markers and non-markers.}

\item{foldchange.cutoff}{Fold chagne cutoff.}

\item{p.cutoff}{p value cutoff.}

\item{vip.cutoff}{VIP value cutoff.}

\item{path}{The directory you want to write results.}
}
\value{
volcano plot.
}
\description{
Draw volcano plot.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

